delta time float to double

+++ test_bot/test_bot.cpp
@@ -1,6 +1,6 @@
 #include "test_bot.h"
 
-void test_actions(float delta) {}
+void test_actions(double delta) {}
 
 void test_init() {
